Usable update But cia still not working

+++ source/stringextra.hpp
@@ -18,7 +18,7 @@ inline std::string last_lines(const std::string &input, int count = 1) {
   //std::reverse(lines.begin(), lines.end());
   line += "Lines Count -> " + std::to_string(lines.size() + 1) + "\n";
 
-  for (size_t i = lines.size()-1-count; i < count; i++) {
+  for (size_t i = lines.size()-1-(size_t)count; i < (size_t)count; i++) {
     line += /*std::to_string(i + 1) + "->" + */lines[i] + "\n";
   }
   return line;
